requirements.txt
aiohttp>=3.9.0
aiosqlite>=0.20.0
fastapi>=0.115.0
pydantic>=2.4.1
python-dotenv>=1.2.0
uvicorn>=0.32.0
yookassa>=3.9.0
